requirements.txt
numpy>=1.23.1
torch>=1.13.0
opencv-python>=4.5.4.60
matplotlib>=3.5.2
torchvision>=0.14.0
Pillow>=9.2.0
datasets>=2.6.1
tqdm>=4.62.3